Docs: Refactored Gestalt Docs Sidebar Menu #952: comment fixes
